fix: [Refactoring] added the services layer to projects
