+++ src/domain/projects/hooks/api/useGetProjects.ts
@@ -1,0 +1,7 @@
+import getProjects from '@domain/projects/services/getProjects';
+
+const useGetProjects = () => {
+  return getProjects();
+};
+
+export default useGetProjects;
