+++ src/pages/Blog.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactElement } from 'react';
 import Filters from '../components/filters/Filters';
-import CardsPagesBlog from '../domain/blog/CardsPagesBlog';
+import CardsPagesBlog from '../domain/blog/components/CardsPagesBlog';
 import usePostsController from '../domain/blog/hooks/usePostsController';
 import { PostsController, PostsType } from '../utils/MultiCardsIntetrfaces';
